backend/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
websockets==11.0.3
pydantic==2.4.2
pyautogen==0.2.0
openai==1.2.0
